Add code for invulnerability frames and collision damage. Demo in my sandbox.

+++ Assets/Scripts/Weapons/BasicLaserWeapon.cs
@@ -8,7 +8,7 @@ public class BasicLaserWeapon : MonoBehaviour, IWeapon
     public float speed = 15;
     public float timeBetweenShots = 1;
     public float projectileLifetime = 10;
-    public float damage = 2;
+    public int damage = 2;
     private float timer = 0;
     public float forwardOffset = 1.5f;
     public float upwardOffset = 1.5f;
